+++ src/Clock/Timebox.php
@@ -4,16 +4,40 @@ declare(strict_types=1);
 
 namespace Tempest\Clock;
 
-interface Timebox
+use Throwable;
+
+final class Timebox
 {
-    /**
-     * Run callback at least for the given number of microseconds.
-     *
-     * @template TCallbackReturnType
-     *
-     * @param  callable(): TCallbackReturnType  $callable
-     * @param  int  $microseconds
-     * @return TCallbackReturnType
-     */
-    public function run(callable $callable, int $microseconds): mixed;
+    public function __construct(
+        protected Clock $clock,
+    ) {
+    }
+
+    public function run(callable $callable, int $microseconds, bool $returnEarly = false): mixed
+    {
+        $throwable = null;
+        $result = null;
+
+        $start = $this->clock->utime();
+
+        try {
+            $result = $callable();
+        } catch (Throwable $exception) {
+            $throwable = $exception;
+        }
+
+        if ($returnEarly) {
+            return $result;
+        }
+
+        $remaining = $microseconds - ($this->clock->utime() - $start);
+
+        $this->clock->usleep($remaining);
+
+        if ($throwable) {
+            throw $throwable;
+        }
+
+        return $result;
+    }
 }
